Factor out route probing logic
