+++ api/src/probe-routes.ts
@@ -1,3 +1,7 @@
+import fs from "node:fs";
+
+import { getIgnoredPaths, shouldIgnoreFile } from "./lib/utils.js";
+
 let debounceTimeout: NodeJS.Timeout | null = null;
 
 // biome-ignore lint/suspicious/noExplicitAny: Trust me, this is easier
@@ -16,20 +20,51 @@ function debounce<T extends (...args: any[]) => void>(func: T, wait: number) {
 /**
  * Since we are calling the route probe inside a file watcher, we should implement
  * debouncing to avoid spamming the service with requests.
- * 
+ *
  * HACK - Since we are monitoring ts files, we need a short delay to let the code
  *        for the service recompile.
  */
-export const debouncedProbeRoutesWithExponentialBackoff = debounce(
+const debouncedProbeRoutesWithExponentialBackoff = debounce(
   probeRoutesWithExponentialBackoff,
   1500,
 );
+
+export function startRouteProbeWatcher(watchDir: string) {
+  // Fire off an async probe to the service we want to monitor
+  // This will collect information on all routes that the service exposes
+  // Which powers a postman-like UI to ping routes and see responses
+  const serviceTargetArgument = process.env.FPX_SERVICE_TARGET;
+  const probeMaxRetries = 10;
+  const probeDelay = 1000;
+
+  debouncedProbeRoutesWithExponentialBackoff(
+    serviceTargetArgument,
+    probeMaxRetries,
+    probeDelay,
+  );
+
+  const ignoredPaths = getIgnoredPaths();
+
+  fs.watch(watchDir, { recursive: true }, async (eventType, filename) => {
+    if (shouldIgnoreFile(filename, ignoredPaths)) {
+      return;
+    }
+
+    console.debug(`File ${filename} ${eventType}, sending a new probe`);
+
+    debouncedProbeRoutesWithExponentialBackoff(
+      serviceTargetArgument,
+      probeMaxRetries,
+      probeDelay,
+    );
+  });
+}
 
 /**
  * Asynchronously probe the routes of a service with exponential backoff.
  * Makes a request to the service root route with the `X-Fpx-Route-Inspector` header set to `enabled`.
  */
-export async function probeRoutesWithExponentialBackoff(
+async function probeRoutesWithExponentialBackoff(
   serviceArg: string | number | undefined,
   maxRetries: number,
   delay = 1000,
